Add controller, domain, view package and move classes respectively
