+++ src/main/java/step5_racing_car/domain/CarList.java
@@ -1,8 +1,6 @@
-package step3_racing_car;
+package step5_racing_car.domain;
 
-import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import java.util.stream.Stream;
 
 public class CarList {
