+++ src/main/java/step5_racing_car/domain/RaceCalculator.java
@@ -1,4 +1,4 @@
-package step3_racing_car;
+package step5_racing_car.domain;
 
 import java.util.Arrays;
 import java.util.stream.IntStream;
