+++ src/main/java/step5_racing_car/domain/RandomGenerator.java
@@ -1,4 +1,4 @@
-package step3_racing_car;
+package step5_racing_car.domain;
 
 import java.util.Random;
