+++ src/main/java/step5_racing_car/view/InputView.java
@@ -1,4 +1,4 @@
-package step3_racing_car;
+package step5_racing_car.view;
 
 import java.util.Scanner;
