+++ src/test/java/step5_racing_car/RacingGameTest.java
@@ -1,8 +1,9 @@
-package step3_racing_car;
+package step5_racing_car;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step5_racing_car.domain.RacingGame;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
